add organization members service

+++ BookingServer/src/main/java/com/server/organization/application/AddMemberCommand.java
@@ -3,7 +3,7 @@ package com.server.organization.application;
 import com.server.organization.domain.enums.Role;
 
 public record AddMemberCommand(
-        Long organizationId,
-        Long userId,
+        int organizationId,
+        int userId,
         Role role
 ) {}
